requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.11.3
python-dotenv==1.0.1
chromadb==0.6.3
sentence-transformers==3.4.1
pypdf==5.4.0
openai==1.72.0
langchain==0.3.23
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
pytest==8.3.5
